pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "betpredictor"
version = "0.1.0"
description = "MLB betting prediction and data analysis tool"
requires-python = ">=3.8"
dependencies = [
    "attrs>=25.3.0",
    "Automat>=25.4.16",
    "certifi>=2025.6.15",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.2",
    "constantly>=23.10.4",
    "cryptography>=45.0.4",
    "cssselect>=1.3.0",
    "defusedxml>=0.7.1",
    "filelock>=3.18.0",
    "greenlet>=3.2.3",
    "hyperlink>=21.0.0",
    "idna>=3.10",
    "incremental>=24.7.2",
    "itemadapter>=0.11.0",
    "itemloaders>=1.3.2",
    "jmespath>=1.0.1",
    "lxml>=5.4.0",
    "MLB-StatsAPI>=1.7.0",
    "numpy>=2.2.6",
    "packaging>=25.0",
    "pandas>=2.3.0",
    "parsel>=1.10.0",
    "playwright>=1.52.0",
    "Protego>=0.4.0",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycparser>=2.22",
    "PyDispatcher>=2.0.7",
    "pyee>=13.0.0",
    "pyOpenSSL>=25.1.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "queuelib>=1.8.0",
    "requests>=2.32.4",
    "requests-file>=2.1.0",
    "Scrapy>=2.13.2",
    "scrapy-playwright>=0.0.43",
    "service-identity>=24.2.0",
    "setuptools>=80.9.0",
    "six>=1.17.0",
    "tldextract>=5.3.0",
    "Twisted>=25.5.0",
    "typing-extensions>=4.14.0",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "w3lib>=2.3.1",
    "zope.interface>=7.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "data*"]

[tool.setuptools.package-dir]
"" = "."